move deleteLoadFlowResult
